Pull get blob properties
